chore: add rules plugins all platform plugins

+++ products/easysearch/build-release.sh
@@ -83,18 +83,26 @@ for p in ${plugins[@]}; do
     fi
     cp -rf $SRC/plugins/$q/sql-jdbc/build/libs/sql-jdbc-$VERSION.jar $DEST/archive/plugins
   fi
-  cp -rf $SRC/plugins/$q/build/distributions/$p-$VERSION.zip $f
-  sha512sum $f |awk -F'/' '{print $1$NF}' > $f.sha512
 
-  if [[ "$(echo "$PUBLISH_RELEASE" | tr '[:upper:]' '[:lower:]')" == "true" ]]; then
-    if [[ "$(echo "$ONLY_DOCKER" | tr '[:upper:]' '[:lower:]')" == "true" ]]; then
-      echo "Publish Docker <Only> image no need to upload with $p"
-    else
-      echo Upload plugin $p to oss
-      oss upload -c $GITHUB_WORKSPACE/.oss.yml -o -f $f -k $PNAME/stable/plugins/$p
-      oss upload -c $GITHUB_WORKSPACE/.oss.yml -o -f $f.sha512 -k $PNAME/stable/plugins/$p
+  for zip in "${files[@]}"; do
+    if [ -e "$zip" ]; then
+      filename=$(basename "$zip")
+      f="$DEST/plugins/$p/$filename"
+      cp -rf "$zip" "$f"
+      (cd "$(dirname "$f")" && sha512sum "$filename" > "$filename.sha512")
+
+        if [[ "$(echo "$PUBLISH_RELEASE" | tr '[:upper:]' '[:lower:]')" == "true" ]]; then
+          if [[ "$(echo "$ONLY_DOCKER" | tr '[:upper:]' '[:lower:]')" == "true" ]]; then
+            echo "Publish Docker <Only> image no need to upload with $p"
+          else
+            echo Upload plugin $p to oss
+            oss upload -c $GITHUB_WORKSPACE/.oss.yml -o -f $f -k $PNAME/stable/plugins/$p
+            oss upload -c $GITHUB_WORKSPACE/.oss.yml -o -f $f.sha512 -k $PNAME/stable/plugins/$p
+          fi
+        fi
+      fi
     fi
-  fi
+  done
 done
 
 # Refresh snapshot and stable cache page
